refactor: extract download logic to utility to fix sonar duplication
